+++ opentelemetry-otlp/CHANGELOG.md
@@ -1,9 +1,9 @@
 # Changelog
 
 ## vNext
-
 - Bump msrv to 1.75.0.
-
+- `OtlpHttpClient.shutdown` `TonicLogsClient.shutdown`, and `TonicTracesClient.shutdown` now explicitly return a result. The
+    semantics of the method have not changed, but you will have a new lint encouraging you to consume these results.
 
 ## 0.27.0